pyproject.toml
[tool.poetry]
name = "querystar"
version = "0.3.3"
description = "Python framework for workflow automations."
authors = ["Adithya Krishnan <krishsandeep@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://querystar.io"
repository = "https://github.com/modelstar-labs/querystar"
packages = [{include = "querystar", from = "src"}]

[tool.poetry.scripts]
querystar = "querystar.cli:main"
qs = "querystar.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.7"
websockets = "^11.0.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
